refactor(practice-template): rename and reorganize deliverable DTOs
