+++ src/practice-template/dtos/create-practice-template-deliverable-request.dto.ts
@@ -1,0 +1,15 @@
+import { IsMongoId, IsNotEmpty, IsNumber } from "class-validator";
+
+export class CreatePracticeTemplateDeliverableRequestDto {
+    @IsMongoId()
+    template: string;
+
+    @IsNotEmpty()
+    title: string;
+
+    description?: string;
+
+    @IsNotEmpty()
+    @IsNumber()
+    estimatedDueOffsetDays: number;
+}
