--- src/practice-template/dtos/practice-deliverable-response.dto.ts
@@ -1,5 +1,0 @@
-export class PracticeTemplateDeliverableResponseDto {
-    title: string;
-    description?: string;
-    estimatedDueOffsetDays: number;
-}
