+++ src/practice-template/dtos/practice-template-deliverable-filter.dto.ts
@@ -1,0 +1,6 @@
+import { PaginationQueryDto } from "@common/dtos/pagination-query.dto";
+
+export class PracticeDeliverableTemplateFilterDto extends PaginationQueryDto {
+    template?: string;
+    title?: string;
+}
