+++ src/practice-template/dtos/practice-template-delliverable-response.dto.ts
@@ -1,0 +1,9 @@
+
+export class PracticeTemplateDeliverableResponseDto {
+  _id: string;
+  title: string;
+  description?: string;
+  estimatedDueOffsetDays: number;
+  createdAt: Date;
+  updatedAt: Date;
+}
