--- src/practice-template/dtos/practice-template-delliverable.dto.ts
@@ -1,14 +1,0 @@
-import { IsInt, IsNotEmpty, IsOptional, IsString } from "class-validator";
-
-export class PracticeTemplateDeliverableDto {
-  @IsString()
-  @IsNotEmpty()
-  title: string;
-
-  @IsOptional()
-  @IsString()
-  description?: string;
-
-  @IsInt()
-  estimatedDueOffsetDays: number;
-}
